........
  r21206 | dirk.petry | 2012-09-26 10:01:55 -0600 (Wed, 26 Sep 2012) | 17 lines
  
  
         New Development: No
              JIRA Issue: CAS-4106
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: this is a test program
    Put in Release Notes: No
               Module(s): gcwrap
             Description: inserted Kumars changes to the pclean calls.
  			They don't work for me, yet, so usepclean 
  			is still false.
  			Made the one plotms call produce a png file.
  
  	
........
  r21218 | dirk.petry | 2012-09-27 05:38:22 -0600 (Thu, 27 Sep 2012) | 16 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: test_hanningsmooth, test_testconcat
    Put in Release Notes: No
               Module(s): gcwrap
             Description: Corrected tasks testconcat and hanningsmooth
  			for the fact that tb.copy() returns a table
  			object even if parameter returnobject is False.
  			(Which would result in memory leaks if not corrected.)
  	
........
  r21247 | dirk.petry | 2012-09-28 09:07:13 -0600 (Fri, 28 Sep 2012) | 15 lines
  
  
         New Development: No
              JIRA Issue: CAS-4113
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: test_uvcontsub, alma-m100-analysis-hpc-regression
    Put in Release Notes: No
               Module(s): gcwrap
             Description: Fixed bug which concerned MMS processing:
  			want_cont was not taken into account when
  			postprocessing the output.
  	
........


git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/stable-2012-09@21256 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ gcwrap/python/scripts/task_uvcontsub.py
@@ -36,25 +36,27 @@ def uvcontsub(vis, field, fitspw, combine, solint, fitorder, spw, want_cont):
         # depends on the time the engines dispatches their sub-MSs
         cont_subMS_list.sort()
         contsub_subMS_list.sort()
-        
-        try:
-            virtualconcat(concatvis=helper._arg['vis'] + ".cont",vis=cont_subMS_list)
-        except Exception, instance:
-            casalog.post("Error concatenating continuum sub-MSs %s: %s" % 
-                         (str(cont_subMS_list),str(instance)),'SEVERE')     
+
+        if want_cont:
+            try:
+                virtualconcat(concatvis=helper._arg['vis'] + ".cont",vis=cont_subMS_list)
+            except Exception, instance:
+                casalog.post("Error concatenating continuum sub-MSs %s: %s" % 
+                             (str(cont_subMS_list),str(instance)),'SEVERE')     
         try:
             virtualconcat(concatvis=helper._arg['vis'] + ".contsub",vis=contsub_subMS_list)
         except Exception, instance:
-            casalog.post("Error concatenating continuum sub-MSs %s: %s" % 
+            casalog.post("Error concatenating continuum-subtracted sub-MSs %s: %s" % 
                          (str(contsub_subMS_list),str(instance)),'SEVERE')     
                 
         # Remove continuum subtraction-SubMSs
-        for subMS in cont_subMS_list:
-            if (os.system("rm -rf " + subMS ) != 0):
-                casalog.post("Problem removing continuum sub-MS %s into working directory" % (subMS),'SEVERE')
+        if want_cont:
+            for subMS in cont_subMS_list:
+                if (os.system("rm -rf " + subMS ) != 0):
+                    casalog.post("Problem removing continuum sub-MS %s into working directory" % (subMS),'SEVERE')
         for subMS in contsub_subMS_list:
             if (os.system("rm -rf " + subMS ) != 0):
-                casalog.post("Problem removing continuum sub-MS %s into working directory" % (subMS),'SEVERE')
+                casalog.post("Problem removing continuum-subtracted sub-MS %s into working directory" % (subMS),'SEVERE')
         
         return True
